design: 동행글 상세화면 상단바 배경색 수정 및 파일 이동
- 동행글 상세화면 상단바 투명색으로 지정
- 특정 화면에서만 쓰이는 상단바라서 home모듈 ui/component 폴더로 이동

Fixes: #36
